requirements.txt
dependency-injector
flask
zipfile
Werkzeug
pillow